+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.0.6 (2019-09-17)
 * `bundled_project` are now loaded quietly (meaning, they are loaded, but not included into the projection)
 * `MongoQuerySettingsDict.pluck_from()`: now skips the `max_items` key because it does not make sense when inherited
+* Projections: now allow using a string of field names, separated by whitespace. Example: `{project: "name age weight"}`
 * Bugfix: in some cases, the `projection` property returned invalid results
 * Bugfix: when the `join` operation includes two relationships that are LEFT JOINed, the query is not broken anymore.
 * Bugfix: compilation of dialect-specific clauses used to fail with some JOINs